requirements.txt
pydantic==2.4.2
pydantic-core==2.10.1
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
streamlit==1.28.2
openai==1.3.0
python-multipart==0.0.6
beautifulsoup4==4.12.2
requests==2.31.0
python-docx==0.8.11
PyPDF2==3.0.1 